Move components into folders with tests in same folders
