+++ src/components/Progress/Progress.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 class Progress extends React.Component {
     render() {
       return (
@@ -10,3 +11,4 @@ class Progress extends React.Component {
   }
 
   export default Progress;
+
