+++ src/components/TodoElement/TodoElement.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class TodoElement extends React.Component {
+    render() {
+        return "done";
+    }
+}
+
+export default TodoElement;
