+++ src/components/index.js
@@ -1,0 +1,8 @@
+export { default as App } from "./App/App";
+export { default as AddNewTodo } from "./AddNewTodo/AddNewTodo";
+export { default as Date } from "./Date/Date";
+export { default as HeaderImage } from "./HeaderImage/HeaderImage";
+export { default as Image } from "./Image/Image";
+export { default as Progress } from "./Progress/Progress";
+export { default as TodoElement } from "./TodoElement/TodoElement";
+export { default as TodoList } from "./TodoList/TodoList";
